requirements.txt
asgiref==3.8.1
Django==5.2.3
django-otp==1.6.0
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
